move setup to service
